+++ alacritty_terminal/src/tty/mod.rs
@@ -33,6 +33,13 @@ pub struct Options {
 
     /// Extra environment variables.
     pub env: HashMap<String, String>,
+
+    /// Specifies whether the Windows shell arguments should be escaped.
+    ///
+    /// - When `true`: Arguments will be escaped according to the standard C runtime rules.
+    /// - When `false`: Arguments will be passed raw without additional escaping.
+    #[cfg(target_os = "windows")]
+    pub escape_args: bool,
 }
 
 /// Shell options.