+++ alacritty_terminal/CHANGELOG.md
@@ -10,6 +10,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## 0.25.1-dev
 
+### Added
+
+- New `escape_args` field on `tty::Options` for Windows shell argument escaping control
+
 ### Changed
 
 - Pass `-q` to `login` on macOS if `~/.hushlogin` is present